lecture-transcribe/requirements.txt
# lecture-transcribe v3 依賴
# 主力/備援引擎由 install_whisperx.sh 管理

notion-client>=2.2.0
pyyaml>=6.0
huggingface-hub>=0.20.0
openpyxl>=3.1.2
python-docx>=1.1.2
anthropic>=0.37.1

# install_whisperx.sh 鎖定版本:
# mlx-whisper
# torch==2.2.2
# torchaudio==2.2.2
# whisperx==3.1.1
# pyannote.audio==3.1.1
